todo_backend/requirements.txt
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2==2.9.7
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
django-cors-headers==4.2.0